extra app modules (#596)
* toolbar module

* lazy load Settings

* fix build

* remove duplicate build for e2e

* Revert "remove duplicate build for e2e"

This reverts commit 94b872239b91ba5c6e25217673768d665ad67de0.
